third_party/blink/web_tests/http/tests/clipboard/async-write-blobs-read-blobs.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
  Async Clipboard write blobs -> read blobs tests
</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../resources/permissions-helper.js"></script>

<script>
async function loadBlob(fileName) {
  const fetched = await fetch(fileName);
  return await fetched.blob();
}

promise_test(async t => {
  await PermissionsHelper.setPermission('clipboard-read', 'granted');
  await PermissionsHelper.setPermission('clipboard-write', 'granted');

  const blobText = new Blob(['test text'], {type: 'text/plain'});
  const blobImage = await loadBlob('resources/greenbox.png');

  assert_equals(blobText.type, 'text/plain');
  assert_equals(blobImage.type, 'image/png');

  await navigator.clipboard.write(
      {'text/plain' : blobText, 'image/png' : blobImage});
  const output = await navigator.clipboard.read();

  assert_equals(Object.keys(output).length, 2);
  assert_equals(output['text/plain'].type, 'text/plain');
  assert_equals(output['image/png'].type, 'image/png');
}, 'Verify write and read clipboard (multiple blobs)');
</script>
